# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Azure AI Chat API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatRequest {
  /** The user's message */
  message: string;
  /** Optional thread ID to continue a conversation */
  threadId?: string;
}

export interface ChatResponse {
  /** The agent's response */
  reply: string;
  /** The thread ID for continuing the conversation */
  threadId: string;
}

export interface ThreadResponse {
  /** The newly created thread ID */
  threadId: string;
}

export interface ErrorResponse {
  /** Error message */
  error: string;
}