# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Azure AI Chat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  /** The user's message */
  message: string;
  /** Optional thread ID to continue a conversation */
  threadId?: string;
}