# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Azure AI Chat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  /** The agent's response */
  reply: string;
  /** The thread ID for continuing the conversation */
  threadId: string;
}