# lib/api-zod/src/generated/types/threadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Azure AI Chat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ThreadResponse {
  /** The newly created thread ID */
  threadId: string;
}